requirements.txt
# Core LLM Framework
langchain>=0.1.0
langgraph>=0.0.40
langsmith>=0.0.70

# Claude API
anthropic>=0.18.0

# Web Framework  
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0

# Environment Management
python-dotenv>=1.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Utility
pydantic>=2.5.0
aiofiles>=23.0.0
nest-asyncio>=1.5.0

# Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Jupyter Notebook Support
jupyter>=1.0.0
jupyterlab>=4.0.0
jupyterlab-language-pack-ja-JP>=4.0.post0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0
japanize-matplotlib>=1.1.3

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Additional Utilities
requests>=2.31.0
typing-extensions>=4.7.0
asyncio>=3.4.3